rename test classes and reorganize package structure
