+++ application/build.gradle
@@ -19,7 +19,6 @@ dependencies {
     // ---- Spring Boot Core ----
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
-    implementation 'org.springframework.boot:spring-boot-starter-aop'
 
     // ---- Data & Database ----
     implementation 'com.mysql:mysql-connector-j:8.3.0'
