+++ application/src/main/java/org/mandarin/booking/app/QuerydslConfig.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class QuerydslConfig {
+class QuerydslConfig {
     @PersistenceContext
     private EntityManager entityManager;
