+++ internal/build.gradle
@@ -10,4 +10,5 @@ dependencies {
     api 'io.jsonwebtoken:jjwt-api:0.12.6'
     api 'io.jsonwebtoken:jjwt-impl:0.12.6'
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.6'
+    implementation 'org.springframework.boot:spring-boot-starter-aop'
 }
